+++ xclim/ensembles/__init__.py
@@ -1,6 +1,4 @@
 """
-Ensembles Module.
-
 This submodule defines some useful methods for dealing with ensembles of climate
 simulations. In xclim, an "ensemble" is a `Dataset` or a `DataArray` where multiple
 climate realizations or models are concatenated along the `realization` dimension.